+++ nrewebservices/__release__.py
@@ -1,1 +1,1 @@
-release = '0.1.4-dev'
+release = '0.1.4'